Fix a big? Optimize? Ignore a bug? One of those

+++ src/camera.rs
@@ -8,10 +8,10 @@ pub struct Camera {
     vertical: Vec3,
 }
 
-const ASPECT_RATIO: f32 = 16. / 9.;
-const VIEWPORT_HEIGHT: f32 = 2.0;
-const VIEWPORT_WIDTH: f32 = ASPECT_RATIO * VIEWPORT_HEIGHT;
-const FOCAL_LENGTH: f32 = 1.0;
+const ASPECT_RATIO: f64 = 16. / 9.;
+const VIEWPORT_HEIGHT: f64 = 2.0;
+const VIEWPORT_WIDTH: f64 = ASPECT_RATIO * VIEWPORT_HEIGHT;
+const FOCAL_LENGTH: f64 = 1.0;
 
 impl Default for Camera {
     fn default() -> Self {
@@ -32,7 +32,7 @@ impl Default for Camera {
 }
 
 impl Camera {
-    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
         Ray::new(
             self.origin,
             self.lower_left_corner + u * self.horizontal + v * self.vertical - self.origin,
